+++ snippet/dot/dot_nxpp.cpp
@@ -1,0 +1,11 @@
+#include <cstdlib>
+#include <iostream>
+
+#include "../../include/nxpp.hpp"
+
+int main() {
+    nxpp::DiGraphInt g;
+    g.add_edge(0, 1, 2);
+    std::cout << g.to_dot_string();
+    return EXIT_SUCCESS;
+}